fix: add missing fields to DatasetInfo model and verify all knowledge base examples
- Added doc_metadata and built_in_field_enabled fields to DatasetInfo model
- Added corresponding builder methods for new fields
- Verified all knowledge base examples work correctly with local Dify instance
- Fixed data structure compatibility with API responses
- All examples now run successfully except retrieve (expected due to local config)

Tested examples:
- Dataset: create, list, get, update, delete (✓)
- Document: create_by_text, create_by_file, list, get, get_upload_file (✓)
- Metadata: create, list, toggle_builtin (✓)
- Tag: create, list, bind, query_bound (✓)
- Cleanup operations: delete dataset, tag, metadata (✓)

Note: Dataset retrieve fails due to missing embedding model in local Dify setup,
which is expected and not an SDK issue.

+++ dify_oapi/api/knowledge_base/v1/model/document/upload_file_info.py
@@ -14,7 +14,7 @@ class UploadFileInfo(BaseModel):
     download_url: str | None = None
     mime_type: str | None = None
     created_by: str | None = None
-    created_at: int | None = None
+    created_at: float | None = None
 
     @staticmethod
     def builder() -> UploadFileInfoBuilder:
@@ -60,6 +60,6 @@ class UploadFileInfoBuilder:
         self._upload_file_info.created_by = created_by
         return self
 
-    def created_at(self, created_at: int) -> UploadFileInfoBuilder:
+    def created_at(self, created_at: float) -> UploadFileInfoBuilder:
         self._upload_file_info.created_at = created_at
         return self
